# server/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/pet/order/list": {
      "get": {
        "operationId": "OrderController_findAll",
        "summary": "订单列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderNo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "业务-订单管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/pet/order/{id}": {
      "get": {
        "operationId": "OrderController_findOne",
        "summary": "订单详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "业务-订单管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/pet/order/{id}/ship": {
      "put": {
        "operationId": "OrderController_ship",
        "summary": "订单发货",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "业务-订单管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/pet/order/{id}/status": {
      "put": {
        "operationId": "OrderController_updateStatus",
        "summary": "更新订单状态",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "业务-订单管理"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/pet/payment/list": {
      "get": {
        "operationId": "PaymentController_findAll",
        "summary": "支付记录列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderNo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "method",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "业务-支付记录"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/app/order/cart/list": {
      "get": {
        "operationId": "AppOrderController_getCart",
        "summary": "购物车列表",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "小程序-购物车与订单"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/app/order/cart": {
      "post": {
        "operationId": "AppOrderController_addCart",
        "summary": "加入购物车",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "小程序-购物车与订单"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      },
      "put": {
        "operationId": "AppOrderController_updateCart",
        "summary": "更新购物车",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "小程序-购物车与订单"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/app/order/cart/remove": {
      "put": {
        "operationId": "AppOrderController_removeCart",
        "summary": "删除购物车",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "小程序-购物车与订单"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/app/order": {
      "post": {
        "operationId": "AppOrderController_create",
        "summary": "创建订单",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "小程序-购物车与订单"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/app/order/pay": {
      "post": {
        "operationId": "AppOrderController_pay",
        "summary": "支付订单",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "小程序-购物车与订单"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/app/order/payNotify": {
      "post": {
        "operationId": "AppOrderController_payNotify",
        "summary": "支付回调(模拟)",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "小程序-购物车与订单"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/app/order/list": {
      "get": {
        "operationId": "AppOrderController_getList",
        "summary": "我的订单列表",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "小程序-购物车与订单"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/app/order/{id}": {
      "get": {
        "operationId": "AppOrderController_findOne",
        "summary": "订单详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "小程序-购物车与订单"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/app/order/{id}/receive": {
      "put": {
        "operationId": "AppOrderController_receive",
        "summary": "确认收货",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "小程序-购物车与订单"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/app/order/{id}/cancel": {
      "put": {
        "operationId": "AppOrderController_cancel",
        "summary": "取消订单",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "小程序-购物车与订单"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
    "/pet/meal/order/{id}": {
      "get": {
        "operationId": "MealOrderController_detail",
        "summary": "配餐订单详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "业务-配餐订单"
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest-Admin",
    "description": "Nest-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
...
      "ShipOrderDto": {
        "type": "object",
        "properties": {
          "logisticsCompany": {
            "type": "string"
          },
          "logisticsNo": {
            "type": "string"
          }
        },
        "required": [
          "logisticsCompany",
          "logisticsNo"
        ]
      },
...
    }
  }
}